12202 edit Labrotary: renamed to LaboratoryPageContent and moved to another dir
